js doesn't like 64-bit numbers

+++ backend/esserver/src/main/java/app/database/DatabaseHelper.java
@@ -17,7 +17,7 @@ public class DatabaseHelper {
         if (index < 0) {
             permissions.canCreate();
             // Creating a new product
-            long productId = (long) (Math.random() * Long.MAX_VALUE);// get this from auto increment in database
+            int productId = (int) (Math.random() * Integer.MAX_VALUE);// get this from auto increment in database
             product.setId(productId);
             Product.list.add(product);
             // TODO insert into database
@@ -29,12 +29,13 @@ public class DatabaseHelper {
         }
     }
 
-    public static void deleteProduct(Long id) {
+    public static void deleteProduct(Integer id) {
         if (id == null) {
             throw new BadRequestException("Invalid product ID");
         }
         int index = IntStream.range(0, Product.list.size()).filter(i -> Product.list.get(i).getId() == id).findFirst().orElseThrow(() -> new NotFoundException("No product found for such ID"));
         Product.list.remove(index);
+        // TODO remove database
     }
 
     public static void deleteUser(UUID id) {
